Update kernel inport source from lenovo_s8-50_osc_201605.tar.bz2

+++ include/uapi/linux/netfilter_ipv4/ipt_ULOG.h
@@ -33,13 +33,13 @@ struct ipt_ulog_info {
 
 /* Format of the ULOG packets passed through netlink */
 typedef struct ulog_packet_msg {
-	unsigned long mark;
-	long timestamp_sec;
-	long timestamp_usec;
+	unsigned int mark;
+	int timestamp_sec;
+	int timestamp_usec;
 	unsigned int hook;
 	char indev_name[IFNAMSIZ];
 	char outdev_name[IFNAMSIZ];
-	size_t data_len;
+	int data_len;
 	char prefix[ULOG_PREFIX_LEN];
 	unsigned char mac_len;
 	unsigned char mac[ULOG_MAC_LEN];
